all proped up and working.
