+++ hdmi2usb/install/bin/atlys-off-on.sh
@@ -1,0 +1,6 @@
+#!/bin/bash -x
+
+atlys-off.sh
+sleep 3
+atlys-on.sh
+
