+++ hdmi2usb/install/bin/atlys-off.sh
@@ -1,0 +1,1 @@
+relay.sh 2 on
