+++ hdmi2usb/install/bin/atlys-on.sh
@@ -1,0 +1,1 @@
+relay.sh 2 off
