+++ hdmi2usb/install/bin/opsis-off-on.sh
@@ -1,0 +1,6 @@
+#!/bin/bash -x
+
+opsis-off.sh
+sleep 3
+opsis-on.sh
+
