+++ hdmi2usb/install/bin/opsis-off.sh
@@ -1,0 +1,1 @@
+relay.sh 0 on
