+++ hdmi2usb/install/bin/relay-read.sh
@@ -1,0 +1,9 @@
+#!/binbash
+
+echo relay 0 is Opsis
+echo relay 2 is Atlys
+echo relay on = board power off
+echo relay off = board power on
+
+relay.sh 0 read
+relay.sh 2 read
