+++ hdmi2usb/install/udev/61-numato-relay-udev.rules
