+++ hdmi2usb/tests/top_turtle.sh
@@ -1,0 +1,2 @@
+./one_htmi2usb_board.sh opsis
+./one_htmi2usb_board.sh atlys
